stock-news-grabber/requirements.txt
newsapi-python
python-dotenv
black
flake8
